+++ connector/get_price.go
@@ -2,6 +2,7 @@ package connector
 
 import (
 	"net/url"
+	"strconv"
 
 	"github.com/pkg/errors"
 
@@ -12,7 +13,7 @@ const (
 	priceEndpoint = "pricer/price"
 )
 
-func (c *Connector) GetPrice(platform, contract string) (models.PriceResponse, error) {
+func (c *Connector) GetPrice(platform, contract string, chainId int64) (models.PriceResponse, error) {
 	var response models.PriceResponse
 
 	parsedUrl, err := url.Parse(priceEndpoint)
@@ -23,6 +24,7 @@ func (c *Connector) GetPrice(platform, contract string) (models.PriceResponse, e
 	query := parsedUrl.Query()
 	query.Set("platform", platform)
 	query.Set("contract", contract)
+	query.Set("chain_id", strconv.FormatInt(chainId, 10))
 	parsedUrl.RawQuery = query.Encode()
 
 	err = c.Get(parsedUrl, &response)