+++ connector/models/price.go
@@ -3,3 +3,4 @@ package models
 import "gitlab.com/tokend/nft-books/price-svc/resources"
 
 type PriceResponse resources.PriceResponse
+type NftPriceResponse resources.NftPriceResponse