+++ resources/model_nft_price_attributes.go
@@ -1,0 +1,12 @@
+/*
+ * GENERATED. Do not modify. Your changes might be overwritten!
+ */
+
+package resources
+
+type NftPriceAttributes struct {
+	// price of NFT in native currency
+	NativeCurrency float32 `json:"native_currency"`
+	// price of NFT in $
+	Usd float32 `json:"usd"`
+}