+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2023-02-23
+
+### Added
+- NFT exchange
+- Mocked NFTs: in case one needs to test service locally, one can simply specify a map in the config of an actual test address and corresponding real NFTs that is valid when using coingecko API.
+- Support for various networks
+- Mocked platforms: in case you need to work with network which does not support by coingecko
+
 ## [1.0.0] - 2022-12-09
  
 ### Added
@@ -12,5 +20,9 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - First relatively stable version of a pricer service that returns a token price using coingecko API.    
+
+[1.1.0]: https://gitlab.com/tokend/nft-books/price-svc/compare/1.0.0...1.1.0
+[1.0.0]: https://gitlab.com/tokend/nft-books/price-svc/compare/1.0.0-rc.0...1.0.0
+[1.0.0-rc.0]: https://gitlab.com/tokend/nft-books/price-svc/tags/1.0.0-rc.0